% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{beta_given_delta}
\alias{beta_given_delta}
\title{beta_given_delta}
\usage{
beta_given_delta(delta, c_vec, sigma = 1, tau = 1, tol = 1e-06,
  slope_baseline = 0.5, slope_treatment = 1)
}
\arguments{
\item{delta}{the standardized mean difference effect size}

\item{sigma}{the true within-case variance}

\item{tau}{the true between-case variance}

\item{tol}{absolute difference between true and estimated delta required for convergence}

\item{slope_baseline}{linear trend during baseline subphase}

\item{slope_treatment}{linear trend during treatment subphase}

\item{c_vector}{vector of multipliers for beta vector}
}
\value{
list containing an estimated beta vector and loss values obtained during convergence
}
\description{
derive a beta vector for an (AB)^k SCD study with desired treatment effect equal to delta. Uses
             gradient descent for convergence
}

