% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single-case-matrix-functions.R
\name{beta_vector}
\alias{beta_vector}
\title{beta_vector}
\usage{
beta_vector(X, y, Sigma)
}
\arguments{
\item{X}{the (4k, N_i)-dimensional design matrix for case i.}

\item{y}{the vector of N_i outcomes for case i}

\item{Sigma}{the AR(1) covariance matrix}
}
\value{
a (4k, 1) vector of beta coefficients
}
\description{
construct the 4k x 1 beta vector corresponding to treatment effect and slope coefficients per subphase
}
\examples{
#assuming you've run >sim <- simulate_ABk(...)
design_1 <- design_matrix_ABk(sim$df[sim$df$case == 1, "treatment"], k = 2)
y_1 <- sim$df[sim$df$case == 1, "outcome"]
sigma_1 <- AR1_matrix(phi = 0.4, rho = 0.05, times = 1:14)
beta_vec <- beta_vector(design_1, y_1, sigma_1)
#NOTE: beta_vec is [intercept_1, intercept_2,...,intercept_k, slope_1, slope_2,...,slope_k], and should match corresp values in sim$betas
}

