% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{c_vector}
\alias{c_vector}
\title{c_vector}
\usage{
c_vector(trt_vec, k, x_is = NULL, treatment_name = "TREATMENT")
}
\arguments{
\item{trt_vec}{character vector indicating sub-phase of a particular case, e.g.
c("BASELINE", "BASELINE", "TREATMENT", "TREATMENT")}

\item{k}{number of complete sub-phase cycles, as in (AB)^k}

\item{x_is}{(2k-1)x1 vector of times at which we evaluate the treatment effect, perhaps subphase introduction times or midpoints of subphases.}

\item{treatment_name}{string indicating name of the treatment sub-phase as recorded in df.}
}
\value{
vector of -1's, 1's, and f(midpoints)
}
\description{
construct c vector containing multipliers and midpoints for multiplying with beta matrix.
       user should make one of these per case.
}
\examples{
c_1 <- c_vector(sim_df[sim_df$case == 1, "treatment"], k = 2)
}

