% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-functions.R
\name{design_matrix_ABk}
\alias{design_matrix_ABk}
\title{design_matrix_ABk}
\usage{
design_matrix_ABk(trt_vec, k, treatment_name = "TREATMENT")
}
\arguments{
\item{trt_vec}{character vector indicating sub-phase of a particular case, e.g.
c("BASELINE", "BASELINE", "TREATMENT", "TREATMENT")}

\item{k}{number of complete sub-phase cycles, as in (AB)^k}

\item{treatment_name}{string indicating name of the treatment sub-phase as recorded in df.}
}
\value{
a (N_i, 4k)-dimension design matrix
}
\description{
construct design matrix for given a given case with k complete baseline-treatment phases.
}
\examples{
X_1 <- design_matrix_ABk(sim_df[sim_df$case == 1, "treatment"], k = 2)
}

