% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_multi_case.R
\name{estimate_effect_size}
\alias{estimate_effect_size}
\title{estimate_effect_size}
\usage{
estimate_effect_size(data, k, rho, phi)
}
\arguments{
\item{data}{data.frame containing case, time, treatment, phase, and outcome columns}

\item{k}{number of complete sub-phase cycles, as in (AB)^k}

\item{rho}{(tau^2)/(tau^2 + sigma^2)}

\item{phi}{first-order autocorrelation parameter}
}
\value{
d, the estimate of the effect size, delta_{R}
}
\description{
calculate d, the estimated effect size when rho and phi are KNOWN.
}
\examples{
beta_matrix <- matrix(rep(c(1:8), 3), byrow = T, nrow = 3)
sim_data <- simulate_ABk(m = 3, n = 20, k = 2, min_Ni = 10, outcome = T, beta_matrix = beta_matrix)
d <- estimate_effect_size(sim_data, 2, 0.5, 0.2)
}

