% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single-case-matrix-functions.R
\name{gather_midpoints}
\alias{gather_midpoints}
\title{gather_midpoints}
\usage{
gather_midpoints(trt_vec, k, treatment_name = "TREATMENT")
}
\arguments{
\item{trt_vec}{character vector indicating sub-phase of a particular case, e.g.}

\item{k}{number of complete baseline, treatment subphases}

\item{treatment_name}{string contained in simulated data's 'treatment' column}
}
\value{
vector of time midpoints of subphases
}
\description{
from vector of time observation points, determine midpoint of each baseline or treatment subphase
}

