% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_multi_case.R
\name{run_equitime_simulations}
\alias{run_equitime_simulations}
\title{run_equitime_simulations}
\usage{
run_equitime_simulations(N = 1000, m = 6, n = 20, k = 2, delta = 2,
  sigma = 1, tau = 1)
}
\arguments{
\item{N}{number of simulations}

\item{m}{number of cases in each multi-SCD study}

\item{n}{number of time observations for each study}

\item{k}{number of complete baseline/treatment cycles}

\item{delta}{true effect size parameter}

\item{sigma}{within-case variance}

\item{tau}{inter-case variance}
}
\value{
d, var_d, g, var_g, d_param: N-vectors related to effect size estimates and variances,
                                     and d_param is the true delta parameter
}
\description{
calculate d and g estimates and variances for N simulations of an m x n multi-SCD studies.
}
\examples{
sim_1000 <- run_equitime_simulations()
hist(sim_1000$d, col = 'grey80')
abline(h = sim_1000$d_param, col = 'red')
}

