% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_ABk}
\alias{simulate_ABk}
\title{simulate_ABk}
\usage{
simulate_ABk(m = 6, n = 12, k = 1, min_Ni = NULL, outcome = FALSE,
  beta_matrix = NULL, phi = 0.2, sigma = 1, tau = 1)
}
\arguments{
\item{m}{number of cases}

\item{n}{maximum number of observations in a case}

\item{k}{number of phases, i.e. complete AB cycles. There are 2k sub-phases.}

\item{min_Ni}{minimum allowable number of time observations in a case.}

\item{outcome}{Boolean, does user want simulated outcome data?}

\item{beta_matrix}{m x 4k matrix with linear beta coefficients}

\item{phi}{the autocorrelation coefficient}

\item{sigma}{within-case variance}

\item{tau}{between-case variance}
}
\value{
data.frame with case, time, treatment, phase, and outcome columns
}
\description{
Simulates data from an $AB^{k}$ study; user has option to simulate obvervational data
             when a matrix of beta coefficients has been supplied.
}
\examples{
sim_df <- simulate_ABk(m = 6, n = 12, k = 2)
}

