% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_ABk_outcomes}
\alias{simulate_ABk_outcomes}
\title{simulate_ABk_outcomes}
\usage{
simulate_ABk_outcomes(df, beta_vec, k, phi = 0.2, sigma = 1, tau = 1,
  treatment_name = "TREATMENT")
}
\arguments{
\item{df}{data.frame from simulate_ABk_no_outcome}

\item{beta_vec}{vector of linear coefficients, shared over individuals}

\item{k}{the number of complete phases}

\item{sigma}{within-case variance}

\item{tau}{between-case variance}

\item{treatment_name}{}

\item{slope_treatment}{string indicating name of the treatment sub-phase as recorded in df}
}
\value{
list containing an estimated beta vector and loss values obtained during convergence
}
\description{
generate outcome vectors for each individual case
}

