% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_multi_case.R
\name{simulate_multi_cases}
\alias{simulate_multi_cases}
\title{simulate_multi_cases}
\usage{
simulate_multi_cases(N = 100, m = 10, n = 30, k = 2, phi = 0.2,
  tau = 1, sigma = 1)
}
\arguments{
\item{N}{number of multi-case simulations}

\item{m}{number of cases within a simulation}

\item{k}{order of the study, i.e. (AB)^k}

\item{phi}{first-order autocorrelation}

\item{tau}{the standard deviation of individual effects}

\item{sigma}{the inter-case residual variance}

\item{maximum}{number of observations in a case}
}
\value{
(N, m) matrix of estimated effect sizes, "d"
}
\description{
simulate N SCD studies and estimate N*m standardized mean differences.
}

