% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandbox.R
\name{simulate_treatment_effects}
\alias{simulate_treatment_effects}
\title{simulate_treatment_effects}
\usage{
simulate_treatment_effects(N = 1000, m = 10, n = 30, k = 2, phi = 0.2,
  tau = 1, sigma = 1, plot_treatment_effects = FALSE)
}
\arguments{
\item{N}{number of multi-case simulations}

\item{m}{number of cases within a simulation}

\item{k}{order of the study, i.e. (AB)^k}

\item{phi}{first-order autocorrelation}

\item{tau}{the standard deviation of individual effects}

\item{sigma}{the inter-case residual variance}

\item{maximum}{number of observations in a case}
}
\value{
(N, m) matrix of estimated treatment effects
}
\description{
simulate N between-subjects studies and estimate N*m estimated treatment effects
}

