% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single-case-matrix-functions.R
\name{T_bar_individual}
\alias{T_bar_individual}
\title{T_bar_individual}
\usage{
T_bar_individual(beta, c_i, k)
}
\arguments{
\item{beta}{the (4k, 1) vector of coefficients ordered [intercept1, intercept2,...,interceptk, slope1, slope2,...,slopek]}

\item{c_i}{the (4k, 1) vector of multipliers and midpoints for counting up beta coefficients}

\item{k}{number of complete sub-phase cycles, as in (AB)^k}
}
\value{
the average T estimate for one case
}
\description{
calculate the mean treatment effect for an individual case
}

