% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single-case-matrix-functions.R
\name{var_i}
\alias{var_i}
\title{var_i}
\usage{
var_i(X, y, Sigma, k)
}
\arguments{
\item{X}{the (4k, N_i)-dimensional design matrix for case i.}

\item{y}{the vector of N_i outcomes for case i}

\item{Sigma}{the AR(1) covariance matrix}

\item{k}{number of complete sub-phase cycles, as in (AB)^k}
}
\value{
sigma_sq_i estimate, A_i an (N_i, N_i)-dimensional matrix (latter value necessary for combining treatment effects across individuals)
}
\description{
calculate the variance about the regression line for case i
}
\examples{
#assuming you've run >sim <- simulate_ABk(...)
sigma_calcs <- var_i(design_1, y_1, sigma_1, k = 2)
}

