% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{visualize_case_acf}
\alias{visualize_case_acf}
\title{visualize_case_acf}
\usage{
visualize_case_acf(df, case = 1)
}
\arguments{
\item{df}{similar to output of simulate_AB}

\item{case}{string of identifier of case in df you wish to plot the ACF for.}
}
\description{
plot ACF of a particular AB^k test case
}
\examples{
sim <- simulate_AB(m = 6, n = 12, k = 2, phi = 0.15, tau = 1)
visualize_case_acf(sim$df)
}

