% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{visualize_sim}
\alias{visualize_sim}
\title{visualize_sim}
\usage{
visualize_sim(df, treatment_name = "TREATMENT")
}
\arguments{
\item{df}{similar to output of simulate_AB}

\item{treatment_name}{string indicating name of the treatment sub-phase as recorded in df.}
}
\description{
plot simulated AB^k data with linear trends.
}
\examples{
sim <- simulate_AB(m = 6, n = 12, k = 2, phi = 0.15, intercept_lambda = 15)
visualize_sim(sim$df)
}

