% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_ABk}
\alias{simulate_ABk}
\title{simulate_ABk}
\usage{
simulate_ABk(m = 6, n = 12, k = 1, phi = 0.2, tau = 1,
  min_Ni = NULL, betas = NULL, sigma = 1)
}
\arguments{
\item{m}{number of cases}

\item{n}{maximum number of observations in a case}

\item{k}{number of phases, i.e. complete AB cycles. There are 2k sub-phases.}

\item{phi}{autocorrelation parameter}

\item{min_Ni}{minimum allowable number of time observations in a case.}

\item{betas}{(m, 4k)-dimensional matrix containing ($\beta_{0}$, $\beta_{1}$) parameters per every sub-phase. If NULL, this has defaults.}

\item{sigma}{amount of within-case residual variance}
}
\value{
data.frame with case, time, treatment, phase, and outcome columns
}
\description{
Simulates data from an $AB^{k}$ test with linear trends in baseline and treatment sub-phases.
}
\examples{
simulate_ABk(m = 6, n = 12, k = 2, phi = 0.15, tau = 1)
}

